ENH: Add "Failed" state to distinguish failures from successfully completions
When modelFilter.slicer.py and volumeFilter.slicer.py scripts were not found, `useProcessOutput` raised an exception (because `pickle.loads(processOutput)` failed) and so `onProcessFinished` was never called.
Fixed the issue by calling `onProcessFinished` regardless of an exception was raised or not.
To allow the caller determine if a processing failed, added the `success` flag to the process class. If `success` flag is True then when execution is finished the process state is set to `Completed`, otherwise the state is set to `Failed`.