BUG: avoid crash on exit by disconnecting signals
As [reported on discouse](https://discourse.slicer.org/t/parrallelprocessing-cannot-work-well-on-m1-mac/24477) some instances of `Process` would crash in the destructors.

The stack trace indicated that some PythonQt code was being invoked during the application's C++ destructor even though the python object was already gone.  This is most likely due to the fact that `Process` (a python class) inherits from `QProcess` (a C++ class)
which is not a common pattern is Slicer and this path is apparently not robust, at least not on MacOS with Apple processors.

This change explicitly disconnects from the `QProcess` signals to avoid the issue.